Move engine and synth into audio lib
